Create fifa-18-more-complete-player-datasetREADME.md
In order to open the file, you need to download the proper dataset. In this study, we review some regression models that have been proposed by statisticians and computer scientists. That review will include machine learning methods like Ridge and Lasso Regression Models, Random Forests, and Decision Tree. Then we compare the accuracy of these methods and try to understand which approach is more efficient to predict our data. In our project, we aim to apply and compare machine learning methods to predict the Euro value of each football player. For comparing these methods we calculated MSE for each regression model.